inference/asr/pyproject.toml
[project]
name = "whats-asr"
version = "0.1.0"
description = "ASR (speech-to-text) gRPC service using faster-whisper"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.67.0,<2.0",
    "grpcio-tools>=1.67.0,<2.0",
    "protobuf>=5.28.0,<6.0",
    "faster-whisper>=1.1.0",
    "numpy>=1.26.0,<2.0",
    "sentencepiece>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/asr"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
